requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-oauth-toolkit==1.2.0
djangorestframework==3.9.0
idna==2.8
oauthlib==2.1.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
PySocks==1.6.8
pytz==2018.7
requests==2.21.0
six==1.12.0
twilio==6.23.1
urllib3==1.24.1
sentry-sdk==0.6.9